mdy_feiju/requirements.txt
aiosqlite==0.22.1
alembic==1.18.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arclet-alconna==1.8.43
arclet-alconna-tools==0.7.12
arrow==1.4.0
bbcode==1.1.0
binaryornot==0.4.4
cashews==7.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cookiecutter==2.6.0
cycler==0.12.1
distlib==0.4.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.3
filetype==1.2.0
fonttools==4.61.1
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
ImageHash==4.3.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.9
linkify-it-py==2.0.3
loguru==0.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdit-py-plugins==0.5.0
mdurl==0.1.2
meme_generator==0.2.2
msgpack==1.1.2
multidict==6.7.1
nb-cli==1.7.0
nepattern==0.7.7
nonebot-adapter-onebot==2.4.6
nonebot-plugin-alconna==0.60.4
nonebot-plugin-imageutils==0.1.17
nonebot-plugin-localstore==0.7.4
nonebot-plugin-orm==0.8.3
nonebot-plugin-petpet==0.3.21
nonebot-plugin-uninfo==0.10.1
nonebot-plugin-waiter==0.8.1
nonebot2==2.4.4

noneprompt==0.1.11
nonestorage==0.1.0
numpy==1.26.4
opencv-python-headless==4.11.0.86
packaging==26.0
Pillow==9.5.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyWavelets==1.8.0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.1
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
tarina==0.7.3
text-unidecode==1.3
textual==7.4.0
tomli==2.4.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.36.1
watchfiles==1.1.1
wcwidth==0.5.0
websockets==16.0
yarl==1.22.0
zipp==3.23.0
jieba
wordcloud
